Rebuild with own moose folder for logfiles
Changed everything to accomplish logging of player, mission and round data to each file. Mission will change but register in the same file as long as the server is not restarted. The player information file will be the same and register all connecting players in to the same file as long as the server is not restarted. The round file is recording player times for each mission round and will change to a new filename for each mission.